+++ acts/src/package/event/chat.rs
@@ -46,7 +46,7 @@ impl ActPackageFn for ManualEventPackage {
 
         let mut params = Vars::new();
         if let Some(ref v) = self.0 {
-            params.insert(consts::ACT_INPUT_DATA.to_string(), json!(v));
+            params.insert(consts::ACT_DATA.to_string(), json!(v));
         }
         let ret = rt.start(&workflow, &params)?;