+++ acts/src/package/tests/code.rs
@@ -1,6 +1,8 @@
+use serde_json::json;
+
 use crate::{
     Act, Workflow,
-    utils::{self, test::create_proc_signal},
+    utils::{self, consts, test::create_proc_signal},
 };
 
 #[tokio::test]
@@ -43,7 +45,7 @@ async fn pack_code_get_data() {
             Act::code(
                 r#"
                 let data = $act.data();
-                return data
+                return { data: data.my_value }
             "#,
             )
             .with_id("code1")
@@ -63,7 +65,7 @@ async fn pack_code_get_data() {
             .first()
             .unwrap()
             .outputs()
-            .get::<String>("my_value")
+            .get::<String>(consts::ACT_DATA)
             .unwrap(),
         "abc"
     );
@@ -78,7 +80,8 @@ async fn pack_code_outputs() {
                 return { "my_output": "abc" };
             "#,
             )
-            .with_id("code1"),
+            .with_id("code1")
+            .with_output("my_output", json!(null)),
         )
     });